Split graphics/simulation.rb into body, trail, v, and extensions.
Namespaced all the main classes and fixed tests/demos.

[git-p4: depot-paths = "//src/graphics/dev/": change = 11036]
