+++ lib/graphics/simulation.rb
@@ -2,7 +2,7 @@
 
 require "sdl"
 
-class Simulation
+class Graphics::Simulation
   D2R = Math::PI / 180.0
   R2D = 1 / D2R
 
@@ -279,258 +279,6 @@ class Simulation
   ensure
     self.screen = old_screen
     self.w, self.h = old_w, old_h
-  end
-end
-
-require "matrix"
-
-class V
-  attr_accessor :x, :y
-
-  class << self
-    alias :[] :new
-  end
-
-  def initialize x, y
-    @x = x
-    @y = y
-  end
-
-  ZERO = V[0.0, 0.0]
-  ONE  = V[1.0, 1.0]
-
-  def + v
-    V[x+v.x, y+v.y]
-  end
-
-  def - v
-    V[x-v.x, y-v.y]
-  end
-
-  def * s
-    V[x*s, y*s]
-  end
-
-  def / s
-    V[x/s, y/s]
-  end
-
-  def == other
-    x == other.x && y == other.y
-  end
-
-  def magnitude
-    Math.sqrt(x*x + y*y)
-  end
-
-  def inspect
-    "#{self.class.name}[%.2f, %.2f]" % [x, y]
-  end
-  alias to_s inspect
-end
-
-class Body
-  D2R = Simulation::D2R
-  R2D = Simulation::R2D
-
-  V_ZERO = V::ZERO
-
-  NORMAL = {
-           :north => 270,
-           :south => 90,
-           :east  => 180,
-           :west  => 0,
-           }
-
-  attr_accessor :x, :y, :a, :ga, :m, :w
-
-  def initialize w
-    self.w = w
-
-    self.x, self.y = rand(w.w), rand(w.h)
-    self.a = 0.0
-    self.ga = 0.0
-    self.m = 0.0
-  end
-
-  def inspect
-    "%s(%.2fx%.2f @ %.2f°x%.2f == %p @ %p)" %
-      [self.class, x, y, a, m, position, velocity]
-  end
-
-  def velocity
-    x, y = dx_dy
-    V[x, y]
-  end
-
-  def velocity= o
-    dx, dy = o.x, o.y
-    self.m = Math.sqrt(dx*dx + dy*dy)
-    self.a = Math.atan2(dy, dx) * R2D
-  end
-
-  def position
-    V[x, y]
-  end
-
-  def position= o
-    self.x = o.x
-    self.y = o.y
-  end
-
-  def dx_dy
-    rad = a * D2R
-    dx = Math.cos(rad) * m
-    dy = Math.sin(rad) * m
-    [dx, dy]
-  end
-
-  def m_a
-    [m, a]
-  end
-
-  def turn dir
-    self.a = (a + dir) % 360.0 if dir
-  end
-
-  def move
-    move_by a, m
-  end
-
-  def move_by a, m
-    rad = a * D2R
-    self.x += Math.cos(rad) * m
-    self.y += Math.sin(rad) * m
-  end
-
-  def clip
-    max_h, max_w = w.h, w.w
-
-    if x < 0 then
-      self.x = 0
-      return :west
-    elsif x > max_w then
-      self.x = max_w
-      return :east
-    end
-
-    if y < 0 then
-      self.y = 0
-      return :north
-    elsif y > max_h then
-      self.y = max_h
-      return :south
-    end
-
-    nil
-  end
-
-  def random_angle
-    360 * rand
-  end
-
-  def random_turn deg
-    rand(deg) - (deg/2)
-  end
-
-  def clip_off_wall
-    if wall = clip then
-      normal = NORMAL[wall]
-      self.ga = (normal + random_turn(90)).degrees unless (normal - ga).abs < 45
-    end
-  end
-
-  def bounce
-    max_h, max_w = w.h, w.w
-    normal = nil
-
-    if x < 0 then
-      self.x, normal = 0, 0
-    elsif x > max_w then
-      self.x, normal = max_w, 180
-    end
-
-    if y < 0 then
-      self.y, normal = 0, 90
-    elsif y > max_h then
-      self.y, normal = max_h, 270
-    end
-
-    if normal then
-      self.a = (2 * normal - 180 - a).degrees
-      self.m *= 0.8
-    end
-  end
-
-  def wrap
-    max_h, max_w = w.h, w.w
-
-    self.x = max_w if x < 0
-    self.y = max_h if y < 0
-
-    self.x = 0 if x > max_w
-    self.y = 0 if y > max_h
-  end
-end
-
-class Trail
-  @@c = {}
-
-  attr_accessor :a, :w, :max, :c
-  def initialize w, max, color = "green"
-    self.w = w
-    self.a = []
-    self.max = max
-    unless @@c[color] then
-      @@c[color] ||= (0..99).map { |n| ("%s%02d" % [color, n]).to_sym }.reverse
-    end
-    self.c = @@c[color]
-  end
-
-  def draw
-    m = 100.0 / max
-    a.reverse_each.each_cons(2).with_index do |((x1, y1), (x2, y2)), i|
-      w.line x1, y1, x2, y2, c[(i*m).round] || :black
-    end
-  end
-
-  def << body
-    a << [body.x, body.y]
-    a.shift if a.size > max
-    nil
-  end
-end
-
-class Integer
-  def =~ n # 1 =~ 50 :: 1 in 50 chance
-    rand(n) <= (self - 1)
-  end
-end
-
-class Numeric
-  def close_to? n, delta = 0.01
-    (self - n).abs < delta
-  end
-
-  def degrees
-    (self < 0 ? self + 360 : self) % 360
-  end
-
-  def relative_angle n, max
-    deltaCW = (self - n).degrees
-    deltaCC = (n - self).degrees
-
-    return if deltaCC < 0.1 || deltaCW < 0.1
-
-    if deltaCC.abs < max then
-      deltaCC
-    elsif deltaCW.close_to? 180 then
-      [-max, max].sample
-    elsif deltaCW < deltaCC then
-      -max
-    else
-      max
-    end
   end
 end
